+++ models/Strategy.py
@@ -177,6 +177,21 @@ class Strategy:
         obv_pc: float = 0.0,
         macdltsignal: bool = False,
     ) -> bool:
+        # set to true for verbose debugging
+        debug = False
+
+        if debug:
+            Logger.warning("\n*** isSellTrigger ***\n")
+
+            Logger.warning("-- loss failsafe sell at fibonacci band --")
+            Logger.warning(f"self.app.disableFailsafeFibonacciLow() is False (actual: {self.app.disableFailsafeFibonacciLow()})")
+            Logger.warning(f"self.app.allowSellAtLoss() is True (actual: {self.app.allowSellAtLoss()})")
+            Logger.warning(f"self.app.sellLowerPcnt() is None (actual: {self.app.sellLowerPcnt()})")
+            Logger.warning(f"self.state.fib_low {self.state.fib_low} > 0")
+            Logger.warning(f"self.state.fib_low {self.state.fib_low} >= {float(price)}")
+            Logger.warning(f"(self.app.allowSellAtLoss() is True (actual: {self.app.allowSellAtLoss()}) or margin ({margin}) > 0)")
+            Logger.warning("\n")
+
         # loss failsafe sell at fibonacci band
         if (
             self.app.disableFailsafeFibonacciLow() is False
@@ -184,6 +199,7 @@ class Strategy:
             and self.app.sellLowerPcnt() is None
             and self.state.fib_low > 0
             and self.state.fib_low >= float(price)
+            and (self.app.allowSellAtLoss() or margin > 0)
         ):
             log_text = (
                 "! Loss Failsafe Triggered (Fibonacci Band: "
@@ -200,14 +216,20 @@ class Strategy:
             )
             return True
 
+        if debug:
+            Logger.warning("-- loss failsafe sell at trailing_stop_loss --")
+            Logger.warning(f"self.app.trailingStopLoss() != None (actual: {self.app.trailingStopLoss()})")
+            Logger.warning(f"change_pcnt_high ({change_pcnt_high}) < self.app.trailingStopLoss() ({self.app.trailingStopLoss()})")
+            Logger.warning(f"margin ({margin}) > self.app.trailingStopLossTrigger() ({self.app.trailingStopLossTrigger()})")
+            Logger.warning(f"(self.app.allowSellAtLoss() is True (actual: {self.app.allowSellAtLoss()}) or margin ({margin}) > 0)")
+            Logger.warning("\n")
+
         # loss failsafe sell at trailing_stop_loss
         if (
             self.app.trailingStopLoss() != None
             and change_pcnt_high < self.app.trailingStopLoss()
-            and (
-                self.app.allowSellAtLoss()
-                and margin > self.app.trailingStopLossTrigger()
-            )
+            and margin > self.app.trailingStopLossTrigger()
+            and (self.app.allowSellAtLoss() or margin > 0)
         ):
             log_text = (
                 "! Trailing Stop Loss Triggered (< "
@@ -224,12 +246,22 @@ class Strategy:
             )
             return True
 
+        if debug:
+            Logger.warning("-- loss failsafe sell at sell_lower_pcnt --")
+            Logger.warning(f"self.app.disableFailsafeLowerPcnt() is False (actual: {self.app.disableFailsafeLowerPcnt()})")
+            Logger.warning(f"and self.app.allowSellAtLoss() is True (actual: {self.app.allowSellAtLoss()})")
+            Logger.warning(f"and self.app.sellLowerPcnt() != None (actual: {self.app.sellLowerPcnt()})")
+            Logger.warning(f"and margin ({margin}) < self.app.sellLowerPcnt() ({self.app.sellLowerPcnt()})")
+            Logger.warning(f"(self.app.allowSellAtLoss() is True (actual: {self.app.allowSellAtLoss()}) or margin ({margin}) > 0)")
+            Logger.warning("\n")
+
         # loss failsafe sell at sell_lower_pcnt
         elif (
             self.app.disableFailsafeLowerPcnt() is False
             and self.app.allowSellAtLoss()
             and self.app.sellLowerPcnt() != None
             and margin < self.app.sellLowerPcnt()
+            and (self.app.allowSellAtLoss() or margin > 0)
         ):
             log_text = (
                 "! Loss Failsafe Triggered (< " + str(self.app.sellLowerPcnt()) + "%)"
@@ -244,11 +276,20 @@ class Strategy:
             )
             return True
 
+        if debug:
+            Logger.warning("-- profit bank at sell_upper_pcnt --")
+            Logger.warning(f"self.app.disableProfitbankUpperPcnt() is False (actual: {self.app.disableProfitbankUpperPcnt()})")
+            Logger.warning(f"and self.app.sellUpperPcnt() != None (actual: {self.app.sellUpperPcnt()})")
+            Logger.warning(f"and margin ({margin}) > self.app.sellUpperPcnt() ({self.app.sellUpperPcnt()})")
+            Logger.warning(f"(self.app.allowSellAtLoss() is True (actual: {self.app.allowSellAtLoss()}) or margin ({margin}) > 0)")
+            Logger.warning("\n")
+
         # profit bank at sell_upper_pcnt
         if (
             self.app.disableProfitbankUpperPcnt() is False
             and self.app.sellUpperPcnt() != None
             and margin > self.app.sellUpperPcnt()
+            and (self.app.allowSellAtLoss() or margin > 0)
         ):
             log_text = (
                 "! Profit Bank Triggered (> " + str(self.app.sellUpperPcnt()) + "%)"
@@ -263,30 +304,21 @@ class Strategy:
             )
             return True
 
-        # profit bank when strong reversal detected
-        if (
-            self.app.disableProfitbankReversal() is False
-            and margin > 3
-            and obv_pc < 0
-            and macdltsignal is True
-        ):
-            log_text = "! Profit Bank Triggered (Strong Reversal Detected)"
-            Logger.warning(log_text)
-            self.app.notifyTelegram(
-                self.app.getMarket()
-                + " ("
-                + self.app.printGranularity()
-                + ") "
-                + log_text
-            )
-            return True
+        if debug:
+            Logger.warning("-- profit bank when strong reversal detected --")
+            Logger.warning(f"self.app.sellAtResistance() is True (actual {self.app.sellAtResistance()})")
+            Logger.warning(f"and price ({price}) > 0")
+            Logger.warning(f"and price ({price}) >= price_exit ({price_exit})")
+            Logger.warning(f"(self.app.allowSellAtLoss() is True (actual: {self.app.allowSellAtLoss()}) or margin ({margin}) > 0)")
+            Logger.warning("\n")
 
         # profit bank when strong reversal detected
         if (
             self.app.sellAtResistance() is True
             and margin >= 2
             and price > 0
-            and price != price_exit
+            and price >= price_exit
+            and (self.app.allowSellAtLoss() or margin > 0)
         ):
             log_text = "! Profit Bank Triggered (Selling At Resistance)"
             Logger.warning(log_text)
@@ -303,6 +335,19 @@ class Strategy:
         return False
 
     def isWaitTrigger(self, margin: float = 0.0, goldencross: bool = False):
+        # set to true for verbose debugging
+        debug = False
+
+        if debug and self.state.action != 'WAIT':
+            Logger.warning("\n*** isWaitTrigger ***\n")
+
+        if debug and self.state.action == 'BUY':
+            Logger.warning("-- if bear market and bull only return true to abort buy --")
+            Logger.warning(f"self.state.action == 'BUY' (actual: {self.state.action})")
+            Logger.warning(f"and self.app.disableBullOnly() is True (actual: {self.app.disableBullOnly()})")
+            Logger.warning(f"and goldencross is False (actual: {goldencross})")
+            Logger.warning("\n")
+
         # if bear market and bull only return true to abort buy
         if (
             self.state.action == "BUY"
@@ -312,6 +357,13 @@ class Strategy:
             log_text = "! Ignore Buy Signal (Bear Buy In Bull Only)"
             Logger.warning(log_text)
             return True
+
+        if debug and self.state.action == 'SELL':
+            Logger.warning("-- configuration specifies to not sell at a loss --")
+            Logger.warning(f"self.state.action == 'SELL' (actual: {self.state.action})")
+            Logger.warning(f"and self.app.allowSellAtLoss() is False (actual: {self.app.allowSellAtLoss()})")
+            Logger.warning(f"and margin ({margin}) <= 0")
+            Logger.warning("\n")
 
         # configuration specifies to not sell at a loss
         if (
@@ -323,8 +375,15 @@ class Strategy:
             Logger.warning(log_text)
             return True
 
+        if debug and self.state.action == 'SELL':
+            Logger.warning("-- configuration specifies not to sell within min and max margin percent bounds --")
+            Logger.warning(f"self.state.action == 'SELL' (actual: {self.state.action})")
+            Logger.warning(f"(self.app.nosellminpcnt is not None (actual: {self.app.nosellminpcnt})) and (margin ({margin}) >= self.app.nosellminpcnt ({self.app.nosellminpcnt}))")
+            Logger.warning(f"(self.app.nosellmaxpcnt is not None (actual: {self.app.nosellmaxpcnt})) and (margin ({margin}) <= self.app.nosellmaxpcnt ({self.app.nosellmaxpcnt}))")
+            Logger.warning("\n")
+
         # configuration specifies not to sell within min and max margin percent bounds
-        if (
+        if self.state.action == "SELL" and (
             (self.app.nosellminpcnt is not None) and (margin >= self.app.nosellminpcnt)
         ) and (
             (self.app.nosellmaxpcnt is not None) and (margin <= self.app.nosellmaxpcnt)