+++ CHANGELOG.md
@@ -14,6 +14,16 @@ Upgrade library dependancies (if required):
 
 -- Binance US is missing 'tradeFee' API endpoint, now returns default fee for this url
 
+## [3.2.12] - 2021-08-23
+
+### Added
+
+-- Verbose debug option to isSellTrigger and isWaitTrigger
+
+### Changed
+
+-- Logic in isSellTrigger and isWaitTrigger
+
 ## [3.2.11] - 2021-08-22
 
 ### Changed